Bug fixed.
Bug description: fixData.json not uploaded to AWS after a new test.

Bug reason: test data (eyetrackerData.json and testData.json) backuped after a new test done and before resumen reading windows loaded. But fixData.json was created BY the resume reading window, so the fixData was not been uploaded, because the uploading was done before it was created.